apps/cms/requirements.txt
bandcamp-api
django==6.0.4
django-cors-headers
python-slugify
psycopg2-binary
beautifulsoup4
PyJWT==2.12.*
